@@ -23,8 +23,8 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "league/flysystem": "^1.0",
+        "php": ">=8.0",
+        "league/flysystem": "^3.0",
         "xinningsu/baidu-bos": "^1.0"
     },
     "require-dev": {
