@@ -9,26 +9,23 @@ use Sulao\BaiduBos\Client;
 
 class AdapterTest extends TestCase
 {
-    public function testGetClient()
+    protected function testDir()
     {
-        $this->assertInstanceOf(
-            Client::class,
-            $this->filesystem()->getAdapter()->getClient()
-        );
-    }
-
-    public function testDir()
-    {
-        $this->assertTrue($this->filesystem()->createDir('adapter_dir/'));
+        $this->filesystem()->createDirectory('adapter_dir/');
         $this->assertEmpty($this->filesystem()->listContents('adapter_dir/'));
-        $this->assertTrue($this->filesystem()->deleteDir('adapter_dir/'));
+        $this->filesystem()->deleteDirectory('adapter_dir/');
     }
 
-    public function testException()
+    protected function testException()
     {
         $filesystem = $this->filesystem2();
 
-        $this->assertFalse($filesystem->write('test.txt', 'test'));
+        $exception = null;
+        try {
+            $filesystem->write('test.txt', 'test');
+        } catch (Throwable $exception) {
+        }
+
 
         $stream = fopen('php://temp', 'w+b');
         fputs($stream, 'adapter test');
@@ -49,10 +46,9 @@ class AdapterTest extends TestCase
         $this->assertFalse($this->filesystem3()->getVisibility('test.txt'));
     }
 
-    public function testAdapter()
+    protected function testAdapter()
     {
         $this->addFile();
-        $this->updateFile();
         $this->getFile();
         $this->copyFile();
         $this->renameFile();
@@ -64,56 +60,25 @@ class AdapterTest extends TestCase
 
     protected function addFile()
     {
-        $this->assertTrue(
-            $this->filesystem()->write(
-                'adapter_test.txt',
-                'adapter test',
-                ['request' => ['connect_timeout' => 10]]
-            )
+        $this->filesystem()->write(
+            'adapter_test.txt',
+            'adapter test',
+            ['request' => ['connect_timeout' => 10]]
         );
 
         $stream = fopen('php://temp', 'w+b');
         fputs($stream, 'adapter test2');
         rewind($stream);
-        $this->assertTrue(
-            $this->filesystem()->writeStream('adapter_test2.txt', $stream)
-        );
+        $this->filesystem()->writeStream('adapter_test2.txt', $stream);
 
-        $this->assertTrue(
-            $this->filesystem()->put('adapter_test3.txt', 'adapter test3')
-        );
+        $this->filesystem()->write('adapter_test3.txt', 'adapter test3');
 
         $stream = fopen('php://temp', 'w+b');
         fputs($stream, 'adapter test4');
         rewind($stream);
-        $this->assertTrue($this->filesystem()->putStream(
+        $this->filesystem()->writeStream(
             'adapter_test/adapter_test4.txt',
             $stream
-        ));
-    }
-
-    protected function updateFile()
-    {
-        $this->assertTrue($this->filesystem()->update(
-            'adapter_test2.txt',
-            'adapter test2.'
-        ));
-        $this->assertEquals(
-            'adapter test2.',
-            $this->filesystem()->read('adapter_test2.txt')
-        );
-
-        $stream = fopen('php://temp', 'w+b');
-        fputs($stream, 'adapter test2');
-        rewind($stream);
-        $this->assertTrue($this->filesystem()->updateStream(
-            'adapter_test2.txt',
-            $stream
-        ));
-
-        $this->assertEquals(
-            'adapter test2',
-            $this->filesystem()->read('adapter_test2.txt')
         );
     }
 
@@ -133,78 +98,70 @@ class AdapterTest extends TestCase
 
     protected function copyFile()
     {
-        $this->assertTrue($this->filesystem()->copy(
+        $this->filesystem()->copy(
             'adapter_test.txt',
             'adapter_test5.txt'
-        ));
+        );
 
         $this->assertTrue(
-            $this->filesystem()->has('adapter_test5.txt')
+            $this->filesystem()->fileExists('adapter_test5.txt')
         );
     }
 
     protected function renameFile()
     {
-        $this->assertTrue($this->filesystem()->rename(
+        $this->filesystem()->move(
             'adapter_test5.txt',
             'adapter_test6.txt'
-        ));
+        );
 
         $this->assertTrue(
-            $this->filesystem()->has('adapter_test6.txt')
+            $this->filesystem()->fileExists('adapter_test6.txt')
         );
         $this->assertFalse(
-            $this->filesystem()->has('adapter_test5.txt')
+            $this->filesystem()->fileExists('adapter_test5.txt')
         );
     }
 
     protected function getMeta()
     {
-        $meta = $this->filesystem()->getMetadata('adapter_test.txt');
-        $this->assertNotFalse($meta);
-
-        $this->assertArrayHasKey('path', $meta);
-        $this->assertArrayHasKey('type', $meta);
-        $this->assertArrayHasKey('size', $meta);
-        $this->assertArrayHasKey('timestamp', $meta);
-
         $this->assertTrue(
-            is_int($this->filesystem()->getTimestamp('adapter_test.txt'))
+            is_int($this->filesystem()->lastModified('adapter_test.txt'))
         );
         $this->assertTrue(
-            is_int($this->filesystem()->getSize('adapter_test.txt'))
+            is_int($this->filesystem()->fileSize('adapter_test.txt'))
         );
         $this->assertTrue(
-            is_string($this->filesystem()->getMimetype('adapter_test.txt'))
+            is_string($this->filesystem()->mimeType('adapter_test.txt'))
         );
     }
 
     protected function visibility()
     {
-        $this->assertTrue($this->filesystem()->setVisibility(
+        $this->filesystem()->setVisibility(
             'adapter_test.txt',
-            \League\Flysystem\AdapterInterface::VISIBILITY_PRIVATE
-        ));
-
-        $this->assertEquals(
-            \League\Flysystem\AdapterInterface::VISIBILITY_PRIVATE,
-            $this->filesystem()->getVisibility('adapter_test.txt')
+            'private'
         );
 
-        $this->assertTrue($this->filesystem()->setVisibility(
+        $this->assertEquals(
+            'private',
+            $this->filesystem()->visibility('adapter_test.txt')
+        );
+
+        $this->filesystem()->setVisibility(
             'adapter_test.txt',
-            \League\Flysystem\AdapterInterface::VISIBILITY_PUBLIC
-        ));
+            'public'
+        );
 
         $this->assertEquals(
-            \League\Flysystem\AdapterInterface::VISIBILITY_PUBLIC,
+            'public',
             $this->filesystem()->getVisibility('adapter_test2.txt')
         );
     }
 
     protected function listContents()
     {
-        $result = $this->filesystem()->listContents();
+        $result = $this->filesystem()->listContents('', false);
         $this->assertTrue(is_array($result));
         $contents = array_column($result, 'path');
         $this->assertTrue(in_array('adapter_test.txt', $contents));
@@ -231,18 +188,18 @@ class AdapterTest extends TestCase
 
     protected function deleteFile()
     {
-        $this->assertTrue($this->filesystem()->delete('adapter_test.txt'));
-        $this->assertTrue($this->filesystem()->delete('adapter_test2.txt'));
-        $this->assertTrue($this->filesystem()->delete('adapter_test3.txt'));
-        $this->assertTrue($this->filesystem()->delete('adapter_test6.txt'));
-        $this->assertTrue($this->filesystem()->delete(
+        $this->filesystem()->delete('adapter_test.txt');
+        $this->filesystem()->delete('adapter_test2.txt');
+        $this->filesystem()->delete('adapter_test3.txt');
+        $this->filesystem()->delete('adapter_test6.txt');
+        $this->filesystem()->delete(
             'adapter_test/adapter_test4.txt'
-        ));
+        );
 
-        $this->assertFalse($this->filesystem()->has('adapter_test.txt'));
+        $this->assertFalse($this->filesystem()->fileExists('adapter_test.txt'));
     }
 
-    protected function filesystem()
+    protected function filesystem(): Filesystem
     {
         static $filesystem;
 
@@ -260,7 +217,7 @@ class AdapterTest extends TestCase
         return $filesystem;
     }
 
-    protected function filesystem2()
+    protected function filesystem2(): Filesystem
     {
         static $filesystem;
 
@@ -278,7 +235,7 @@ class AdapterTest extends TestCase
         return $filesystem;
     }
 
-    protected function filesystem3()
+    protected function filesystem3(): Filesystem
     {
         static $filesystem;
 
